feat: #88 標準報酬月額表のdaoレイヤをD1に移動
